Backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pillow>=10.2.0
requests>=2.31.0
numpy>=1.26.0
torch>=2.2.0
transformers>=4.38.0
huggingface_hub>=0.21.0
scipy>=1.12.0
trimesh>=4.2.0
accelerate>=0.27.0
opencv-python-headless>=4.9.0
